optimizing weno

+++ Weno/Ju/Convection.jl
@@ -1,9 +1,9 @@
 module Convection
 export flux,minf,maxf
-function flux(x)
+function flux(x::Float64)
     x
 end
-function minf(a,b)
+function minf(a::Float64,b::Float64)
     if b<=0.0
         return flux(b)
     elseif a>=0.0
@@ -12,7 +12,7 @@ function minf(a,b)
         return 0.0
     end
 end
-function maxf(a,b)
+function maxf(a::Float64,b::Float64)
     max(flux(a),flux(b))
 end
 end
